fix: primary btn and spacing #127
